requirements.txt
telethon>=1.36.0
python-dotenv>=1.0.0
cryptg>=0.4.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
